Demo简化
